Ensure that connection is not closed before response is fully writen
The HttpStreamsHandler is notifying that a FullHttpMessage is fully sent
after it's written and flushed.  The problem is, it may still be in the
Netty buffers, because the OS buffers are full.  If the connection is
then closed (eg, because using HTTP/1.0 with no keep alive), the close
signal will overtake the buffered part of the message, resulting in an
incomplete response.

This ensures that the message sent callback is not invoked until after
the promise associated with the write is complete, which indicates that
the message has been fully flushed to the OS buffers.

Fixes https://github.com/playframework/playframework/issues/5835.