Release reference to HttpStreamsServerHandler.lastRequest when possible
When using idleTimeout configuration in a play application, the
HttpStreamsServerHandler keeps a reference to lastRequest, and
lastRequest can be relatively large (10kb).  When there are many open
idle connections, lastRequest heap usage can add up, even though the
data is not used (lastRequest will be reassigned to a new value when
serving the next request).

This change sets lastRequest to null to allow the HttpRequest to be GC'd
even if HttpStreamsServerHandler is kept in the heap.